[ADD] CREACION DE PUBLICACIONES, ACTIVIDADES

+++ app/src/main/java/com/example/classroom/presentation/screens/activity/addActivity/AddActivityScreen.kt
@@ -231,7 +231,7 @@ fun AddActivityScreen(
                                         statusId = state.status.id,
                                         startDate = currentDateFormatted,
                                         endDate = state.endDate,
-                                        idCourse = idCourse,
+                                        idCourse = idCourse.toInt(),
                                         grade = state.grade.toInt(),
                                         email = email
                                     )
